+++ Blueprints/BlueprintDefinitions/vs2022/MessageProcessingFramework/template/src/BlueprintBaseName.1/.lambda-test-tool/SavedRequests/HandlerSampleRequest.json
