+++ Blueprints/BlueprintDefinitions/vs2022/MessageProcessingFramework/template/src/BlueprintBaseName.1/Functions.cs
@@ -20,9 +20,7 @@ public class Functions
     private ILambdaMessaging _lambdaMessaging;
 
     /// <summary>
-    /// Default constructor. This constructor is used by Lambda to construct the instance. When invoked in a Lambda environment
-    /// the AWS credentials will come from the IAM role associated with the function and the AWS region will be set to the
-    /// region the Lambda function is executed in.
+    /// This constructor is used by Lambda to construct the instance.
     /// </summary>
     /// <param name="lambdaMessaging">Framework component that processes messages in Lambda</param>
     public Functions(ILambdaMessaging lambdaMessaging)
@@ -36,15 +34,16 @@ public class Functions
     /// <param name="publisher">Generic message publisher, can send messages or publish events 
     /// to any of the destinations that are configured in Startup.cs</param>
     /// <param name="message">Message that is received as an input to the Lambda function then forwarded to SQS</param>
+    /// <param name="context">Lambda execution context</param>
     [LambdaFunction(Policies = "AmazonSQSFullAccess")]
-    public async Task Sender([FromServices] IMessagePublisher publisher, [FromBody] GreetingMessage message)
+    public async Task Sender([FromServices] IMessagePublisher publisher, GreetingMessage message, ILambdaContext context)
     {
         if (message == null)
         {
             return;
         }
 
-        Console.WriteLine($"Received {message.Greeting} from {message.SenderName}, will send to SQS");
+        context.Logger.LogLine($"Received {message.Greeting} from {message.SenderName}, will send to SQS");
 
         // Publish the message to the queue configured in Startup.cs
         await publisher.PublishAsync(message);
@@ -62,31 +61,3 @@ public class Functions
         await _lambdaMessaging.ProcessLambdaEventAsync(evnt, context);
     }
 }
-
- /// <summary>
- /// Business logic for processing <see cref="GreetingMessage"/> messages
- /// </summary>
- public class GreetingMessageHandler : IMessageHandler<GreetingMessage>
- {
-     /// <summary>
-     /// This handler will be invoked once for each <see cref="GreetingMessage"/> that is received
-     /// </summary>
-     /// <param name="messageEnvelope">Envelope that wraps the actual message with metadata used by the framework</param>
-     /// <param name="token">Cancellation token</param>
-     /// <returns>The appropriate MessageProcessStatus based on whether the message was processed successfully</returns>
-     public Task<MessageProcessStatus> HandleAsync(MessageEnvelope<GreetingMessage> messageEnvelope, CancellationToken token = default)
-     {
-         // The outer envelope contains metadata, and its Message property contains the actual message content
-         var greetingMessage = messageEnvelope.Message;
-
-         if (string.IsNullOrEmpty(greetingMessage.Greeting) || string.IsNullOrEmpty(greetingMessage.SenderName))
-         {
-             return Task.FromResult(MessageProcessStatus.Failed());
-         }
-
-         Console.WriteLine($"Received '{greetingMessage.Greeting}' from '{greetingMessage.SenderName}'");
-
-         return Task.FromResult(MessageProcessStatus.Success());
-     }
- }
-
